feat: move write blocks for entity
